refactor(api-search-index-tasks-ddb-os): organize files by concern
Group files into storage/, indexManager/, settings/ directories
instead of flat root layout.
