+++ packages/api-search-index-tasks-ddb-os/src/storage/StorageWriter.ts
